+++ crates/matrix-sdk-crypto/CHANGELOG.md
@@ -12,6 +12,9 @@
   keys whose metadata lacked check fields.
   ([#3046](https://github.com/matrix-org/matrix-rust-sdk/pull/3046))
 
+- Add new API `OlmMachine::create_encrypted_to_device_request` that allows
+  to encrypt a to device event to a specific device.
+
 # 0.7.0
 
 - Add method to mark a list of inbound group sessions as backed up: